requirements.txt
qdrant-client==1.15.1
python-dotenv==1.1.1
feedparser==6.0.11
requests~=2.32.4
httpx~=0.28.1